# public/swagger-suno-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "suno-api",
    "description": "Suno API Final runtime, including guarded direct-HTTP Studio adapters and a local CDP Transport bridge for an already loaded Studio page.",
    "version": "",
    "license": {
      "name": "gcui-art",
      "url": "https://github.com/gcui-art/"
    }
  },
...
  "paths": {
    "/api/custom_generate": {
      "post": {
        "summary": "Generate Audio - Custom Mode",
        "description": "The custom mode enables users to provide additional details about the music, such as music genre, lyrics, and more.2 audio files will be generated for each request, consuming a total of 10 credits.wait_audio can be set to API mode:• By default, it is set to false, which indicates the background mode. It will only return audio task information, and you will need to call the get API to retrieve detailed audio information.• If set to true, it simulates synchronous mode. The API will wait for a maximum of 100s until the audio is generated, and will directly return the audio link and other information. Recommend using in GPTs and other agents.",
        "tags": [
          "默认分组"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "prompt",
                  "tags",
                  "title"
                ],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Detailed prompt, including information such as music lyrics.",
                    "example": "[Verse 1]\nCruel flames of war engulf this land\nBattlefields filled with death and dread\nInnocent souls in darkness, they rest\nMy heart trembles in this silent test\n\n[Verse 2]\nPeople weep for loved ones lost\nBattered bodies bear the cost\nSeeking peace and hope once known\nOur grief transforms to hearts of stone\n\n[Chorus]\nSilent battlegrounds, no birds' song\nShadows of war, where we don't belong\nMay flowers of peace bloom in this place\nLet's guard this precious dream with grace\n\n[Bridge]\nThrough the ashes, we will rise\nHand in hand, towards peaceful skies\nNo more sorrow, no more pain\nTogether, we'll break these chains\n\n[Chorus]\nSilent battlegrounds, no birds' song\nShadows of war, where we don't belong\nMay flowers of peace bloom in this place\nLet's guard this precious dream with grace\n\n[Outro]\nIn unity, our strength will grow\nA brighter future, we'll soon know\nFrom the ruins, hope will spring\nA new dawn, we'll together bring"
                  },
                  "tags": {
                    "type": "string",
                    "description": "Music genre",
                    "example": "pop metal male melancholic"
                  },
                  "title": {
                    "type": "string",
                    "description": "Music title",
                    "example": "Silent Battlefield"
                  },
                  "make_instrumental": {
                    "type": "boolean",
                    "description": "Whether to generate instrumental music",
                    "example": "false"
                  },
                  "model": {
                    "type": "string",
                    "description": "Suno V6 model ID: chirp-hawk (V6), chirp-hawk-wild (V6 Wild), or chirp-goose (V6 Mini). Legacy V5 IDs normalize to chirp-hawk.",
                    "example": "chirp-hawk|chirp-hawk-wild|chirp-goose"
                  },
                  "wait_audio": {
                    "type": "boolean",
                    "description": "Whether to wait for music generation, default is false, directly return audio task information; set to true, will wait for up to 100s until the audio is generated.",
                    "example": "false"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "0",
                      "1"
                    ],
                    "properties": [
                      {
                        "$ref": "#/components/schemas/audio info"
                      },
                      {
                        "$ref": "#/components/schemas/audio info"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/generate": {
      "post": {
        "summary": "Generate audio based on Prompt.",
        "description": "It will automatically fill in the lyrics.2 audio files will be generated for each request, consuming a total of 10 credits.wait_audio can be set to API mode:• By default, it is set to false, which indicates the background mode. It will only return audio task information, and you will need to call the get API to retrieve detailed audio information.• If set to true, it simulates synchronous mode. The API will wait for a maximum of 100s until the audio is generated, and will directly return the audio link and other information. Recommend using in GPTs and other agents.",
        "tags": [
          "默认分组"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "prompt",
                  "make_instrumental",
                  "wait_audio"
                ],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "Prompt",
                    "example": "A popular heavy metal song about war, sung by a deep-voiced male singer, slowly and melodiously. The lyrics depict the sorrow of people after the war."
                  },
                  "make_instrumental": {
                    "type": "boolean",
                    "description": "Whether to generate instrumental music",
                    "example": "false"
                  },
                  "model": {
                    "type": "string",
                    "description": "Suno V6 model ID: chirp-hawk (V6), chirp-hawk-wild (V6 Wild), or chirp-goose (V6 Mini). Legacy V5 IDs normalize to chirp-hawk.",
                    "example": "chirp-hawk|chirp-hawk-wild|chirp-goose"
                  },
                  "wait_audio": {
                    "type": "boolean",
                    "description": "Whether to wait for music generation, default is false, directly return audio task information; set to true, will wait for up to 100s until the audio is generated.",
                    "example": "false"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "0",
                      "1"
                    ],
                    "properties": [
                      {
                        "$ref": "#/components/schemas/audio info"
                      },
                      {
                        "$ref": "#/components/schemas/audio info"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}